feat: 调整knowledgerole目录结构,合并主分支修改
