+++ apps/web-antd/src/views/operator/knowledgeRole/knowledgeRoleGroup/index.vue
@@ -2,12 +2,11 @@
 import type { VbenFormProps } from '@vben/common-ui';
 
 import type { VxeGridProps } from '#/adapter/vxe-table';
-import type { PageQuery } from '#/api/common';
-import type { KnowledgeRole } from '#/api/system/knowledgeRole/knowledge-role-model';
+import type { KnowledgeRoleGroup } from '#/api/operator/knowledgeRole/knowledge-role-group-model';
 
 import { ref } from 'vue';
 
-import { useVbenDrawer } from '@vben/common-ui';
+import { useVbenModal } from '@vben/common-ui';
 import { getVxePopupContainer } from '@vben/utils';
 
 import { Modal, Popconfirm, Space } from 'ant-design-vue';
@@ -16,15 +15,12 @@ import { useVbenVxeGrid, vxeCheckboxChecked } from '#/adapter/vxe-table';
 
 import { emitter } from '../mitt';
 import { columns, querySchema } from './data';
-import knowledgeRoleDrawer from './knowledge-role-drawer.vue';
-import { knowledgeRoleList, knowledgeRoleRemove } from '#/api/system/knowledgeRole/knowledge-role';
-import { optionOptions } from 'ant-design-vue/es/vc-mentions/src/Option';
-
-const groupId = ref('');
+import knowledgeRoleGroupModal from './knowledge-role-group-modal.vue';
+import { knowledgeRoleGroupList, knowledgeRoleGroupRemove } from '#/api/operator/knowledgeRole/knowledge-role-group';
 
 const formOptions: VbenFormProps = {
   commonConfig: {
-    labelWidth: 80,
+    labelWidth: 70,
     componentProps: {
       allowClear: true,
     },
@@ -49,97 +45,97 @@ const gridOptions: VxeGridProps = {
   proxyConfig: {
     ajax: {
       query: async ({ page }, formValues = {}) => {
-        const params: PageQuery = {
+        return await knowledgeRoleGroupList({
           pageNum: page.currentPage,
           pageSize: page.pageSize,
           ...formValues,
-        };
-        if (groupId.value) {
-          params.groupId = groupId.value;
-        }
-
-        return await knowledgeRoleList(params);
+        });
       },
     },
   },
   rowConfig: {
     keyField: 'id',
   },
-  id: 'knowledge-role-index',
+  id: 'knowledge-role-group-index',
 };
+
+const lastGroupId = ref('');
 
 const [BasicTable, tableApi] = useVbenVxeGrid({
   formOptions,
   gridOptions,
+  gridEvents: {
+    cellClick: (e) => {
+      const { row } = e;
+      if (lastGroupId.value === row.id) {
+        return;
+      }
+      emitter.emit('rowClick', row.id);
+      lastGroupId.value = row.id;
+    },
+  },
 });
-
-const [KnowledgeRoleDrawer, drawerApi] = useVbenDrawer({
-  connectedComponent: knowledgeRoleDrawer
+const [KnowledgeRoleGroupModal, modalApi] = useVbenModal({
+  connectedComponent: knowledgeRoleGroupModal,
 });
 
 function handleAdd() {
-  drawerApi.setData({ groupId: groupId.value });
-  drawerApi.open();
+  modalApi.setData({});
+  modalApi.open();
 }
 
-async function handleEdit(record: KnowledgeRole) {
-  drawerApi.setData(record);
-  drawerApi.open();
+async function handleEdit(record: KnowledgeRoleGroup) {
+  modalApi.setData({ id: record.id });
+  modalApi.open();
 }
 
-async function handleDelete(row: KnowledgeRole) {
-  await knowledgeRoleRemove([row.id]);
+async function handleDelete(row: KnowledgeRoleGroup) {
+  await knowledgeRoleGroupRemove([row.id]);
   await tableApi.query();
 }
 
 function handleMultiDelete() {
   const rows = tableApi.grid.getCheckboxRecords();
-  const ids = rows.map((row: KnowledgeRole) => row.id);
+  const ids = rows.map((row: KnowledgeRoleGroup) => row.id);
   Modal.confirm({
     title: '提示',
     okType: 'danger',
     content: `确认删除选中的${ids.length}条记录吗？`,
     onOk: async () => {
-      await knowledgeRoleRemove(ids);
+      await knowledgeRoleGroupRemove(ids);
       await tableApi.query();
     },
   });
 }
-
-emitter.on('rowClick', async (value) => {
-  groupId.value = value;
-
-  await tableApi.query();
-});
 </script>
 
 <template>
   <div>
-    <BasicTable id="knowledge-role" table-title="知识库角色列表">
+    <BasicTable id="knowledge-role-group" table-title="知识库角色组">
       <template #toolbar-tools>
         <Space>
           <a-button :disabled="!vxeCheckboxChecked(tableApi)" danger type="primary" @click="handleMultiDelete">
             {{ $t('pages.common.delete') }}
           </a-button>
-          <a-button :disabled="groupId === ''" type="primary" @click="handleAdd">
+          <a-button type="primary" @click="handleAdd">
             {{ $t('pages.common.add') }}
           </a-button>
         </Space>
       </template>
       <template #action="{ row }">
         <Space>
-          <ghost-button @click="handleEdit(row)">
+          <ghost-button @click.stop=" handleEdit(row)">
             {{ $t('pages.common.edit') }}
           </ghost-button>
-          <Popconfirm :get-popup-container="(node) => getVxePopupContainer(node, 'knowledge-role')
+          <Popconfirm :get-popup-container="(node) => getVxePopupContainer(node, 'knowledge-role-group')
             " placement="left" title="确认删除？" @confirm="handleDelete(row)">
-            <ghost-button danger @click.stop="">
+            <ghost-button danger v-access:code="['system:dict:remove']" @click.stop="">
               {{ $t('pages.common.delete') }}
             </ghost-button>
           </Popconfirm>
         </Space>
       </template>
     </BasicTable>
-    <KnowledgeRoleDrawer @reload="tableApi.query()" />
+    <KnowledgeRoleGroupModal @reload="tableApi.query()" />
   </div>
 </template>
